backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
cloudinary==1.38.0
firebase-admin==6.4.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
